requirements.txt
streamlit==1.32.2
PyPDF2==3.0.1
python-dotenv==0.21.0
google-generativeai==0.7.2
langchain-google-genai==1.0.10
langchain==0.2.15
langchain-community==0.2.14
langchain-core==0.2.36
PyMuPDF==1.24.2
fastapi==0.111.0
uvicorn==0.29.0
streamlit-pdf-viewer==0.0.14
pymongo==4.8.0
openai==1.30.5
python-multipart==0.0.9
spacy==3.7.5


